Fix comprehensive linting errors and improve code quality
- Fix gocritic issues: elseif, singleCaseSwitch, assignOp
- Resolve gocyclo high complexity (16+) in 3 functions via Extract Method pattern
- Fix govet shadow variable declarations across multiple files
- Refactor TestAuditLogger into 5 focused subtests for better maintainability
- Refactor LoggingHandler.Handle with helper methods for complexity reduction
- Refactor TestContext into 6 focused subtests with parallel execution capability
- Update type references after PrometheusExporter → Exporter rename
- Maintain backward compatibility and functionality while improving code quality
- All tests pass, code compiles successfully, no breaking changes

+++ pkg/mesh/service.go
@@ -110,7 +110,7 @@ func (s *Service) GetMetrics() resources.ServiceMetrics {
 
 func (s *Service) updateLatencyMetrics(duration time.Duration) {
 	s.metrics.AverageLatency = time.Duration(
-		(int64(s.metrics.AverageLatency)*int64(s.metrics.SuccessfulCalls-1) +
-			int64(duration)) / int64(s.metrics.SuccessfulCalls),
+		(int64(s.metrics.AverageLatency)*(s.metrics.SuccessfulCalls-1) +
+			int64(duration)) / s.metrics.SuccessfulCalls,
 	)
 }
